# Conflicts fixed
#	AudioKit/Common/MIDI/Sequencer/AKSequencer.swift
#	AudioKit/Common/Nodes/Mixing/Mixer/AKMixer.swift
#	AudioKit/Common/Nodes/Playback/Sampler/AKSampler+SoundFont.swift

+++ AudioKit/Common/Internals/Soundpipe/modules/thresh.c
@@ -1,9 +1,9 @@
 /*
  * Foo
- * 
+ *
  * This is a dummy module. It doesn't do much.
  * Feel free to use this as a boilerplate template.
- * 
+ *
  */
 
 #include <stdlib.h>
@@ -40,7 +40,7 @@ int sp_thresh_compute(sp_data *sp, sp_thresh *p, SPFLOAT *in, SPFLOAT *out)
         return SP_OK;
     }
 
-    switch(p->mode) {
+    switch((int)p->mode) {
         /* input signal goes above threshold */
         case 0:
             *out = (*in > p->thresh && p->prev <= p->thresh);
@@ -62,6 +62,6 @@ int sp_thresh_compute(sp_data *sp, sp_thresh *p, SPFLOAT *in, SPFLOAT *out)
     }
 
     p->prev = *in;
-    
+
     return SP_OK;
 }
